move browser testUtils
