+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/customizable-e2e-test/test/browserB2C.spec.ts
@@ -1,13 +1,23 @@
 import * as puppeteer from "puppeteer";
-import { Screenshot, createFolder, setupCredentials, ONE_SECOND_IN_MS } from "e2e-test-utils/src/TestUtils";
+import { 
+    Screenshot, 
+    createFolder, 
+    setupCredentials, 
+    ONE_SECOND_IN_MS,
+    b2cAadPpeEnterCredentials, 
+    b2cLocalAccountEnterCredentials, 
+    clickLoginPopup, 
+    clickLoginRedirect, 
+    waitForReturnToApp,
+    getBrowser, 
+    getHomeUrl 
+} from "e2e-test-utils/src/TestUtils";
 import { BrowserCacheUtils } from "e2e-test-utils/src/BrowserCacheTestUtils";
 import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
 import { AzureEnvironments, AppTypes, UserTypes, B2cProviders } from "e2e-test-utils/src/Constants";
 import { LabClient } from "e2e-test-utils/src/LabClient";
 import { msalConfig as b2cMsalConfig, request as b2cTokenRequest } from "../authConfigs/b2cAuthConfig.json";
-import { b2cAadPpeEnterCredentials, b2cLocalAccountEnterCredentials, clickLoginPopup, clickLoginRedirect, waitForReturnToApp } from "./testUtils";
 import fs from "fs";
-import {getBrowser, getHomeUrl} from "../../testUtils";
 
 const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots/default tests`;
 let sampleHomeUrl = "";
