+++ samples/msal-browser-samples/VanillaJSTestApp2.0/app/pop/test/browser.spec.ts
@@ -1,11 +1,17 @@
 import * as puppeteer from "puppeteer";
-import { Screenshot, createFolder, setupCredentials, enterCredentials } from "e2e-test-utils/src/TestUtils";
+import { 
+    Screenshot, 
+    createFolder, 
+    setupCredentials, 
+    enterCredentials,
+    getBrowser, 
+    getHomeUrl
+} from "e2e-test-utils/src/TestUtils";
 import { BrowserCacheUtils } from "e2e-test-utils/src/BrowserCacheTestUtils";
 import { LabApiQueryParams } from "e2e-test-utils/src/LabApiQueryParams";
 import { AzureEnvironments, AppTypes } from "e2e-test-utils/src/Constants";
 import { LabClient } from "e2e-test-utils/src/LabClient";
 import { JWK, JWT } from "jose";
-import {getBrowser, getHomeUrl} from "../../testUtils";
 
 const SCREENSHOT_BASE_FOLDER_NAME = `${__dirname}/screenshots`;
 let sampleHomeUrl = "";
